+++ src/Flowline/Commands/ProvisionCommand.cs
@@ -46,8 +46,6 @@ public class ProvisionCommand(IAnsiConsole console, FlowlineRuntimeOptions runti
 
     protected override async Task<int> ExecuteFlowlineAsync(CommandContext context, Settings settings, CancellationToken cancellationToken)
     {
-        ValidateForce(context, settings);
-
         // Production URL is required
         var prodEnv = await GetAndCheckEnvironmentInfoAsync(EnvironmentRole.Prod, settings.ProdUrl, settings, cancellationToken);